+++ src/Symfony/Bridge/Twig/Tests/Extension/FormExtensionDivLayoutTest.php
@@ -106,17 +106,10 @@ class FormExtensionDivLayoutTest extends AbstractDivLayoutTest
 
     public function isSelectedChoiceProvider()
     {
-        // The commented cases should not be necessary anymore, because the
-        // choice lists should assure that both values passed here are always
-        // strings
         return array(
-//             array(true, 0, 0),
             array(true, '0', '0'),
             array(true, '1', '1'),
-//             array(true, false, 0),
-//             array(true, true, 1),
             array(true, '', ''),
-//             array(true, null, ''),
             array(true, '1.23', '1.23'),
             array(true, 'foo', 'foo'),
             array(true, 'foo10', 'foo10'),
@@ -140,11 +133,6 @@ class FormExtensionDivLayoutTest extends AbstractDivLayoutTest
     protected function renderForm(FormView $view, array $vars = array())
     {
         return (string) $this->extension->renderer->renderBlock($view, 'form', $vars);
-    }
-
-    protected function renderEnctype(FormView $view)
-    {
-        return (string) $this->extension->renderer->searchAndRenderBlock($view, 'enctype');
     }
 
     protected function renderLabel(FormView $view, $label = null, array $vars = array())