+++ src/Symfony/Bundle/FrameworkBundle/Console/Descriptor/TextDescriptor.php
@@ -35,8 +35,8 @@ class TextDescriptor extends Descriptor
     {
         $showControllers = isset($options['show_controllers']) && $options['show_controllers'];
         $headers = array('Name', 'Method', 'Scheme', 'Host', 'Path');
-        $table = new Table($this->getOutput());
-        $table->setStyle('compact');
+
+        $table = new Table($this->output);
         $table->setHeaders($showControllers ? array_merge($headers, array('Controller')) : $headers);
 
         foreach ($routes->all() as $name => $route) {
@@ -100,8 +100,7 @@ class TextDescriptor extends Descriptor
      */
     protected function describeContainerParameters(ParameterBag $parameters, array $options = array())
     {
-        $table = new Table($this->getOutput());
-        $table->setStyle('compact');
+        $table = new Table($this->output);
         $table->setHeaders(array('Parameter', 'Value'));
 
         foreach ($this->sortParameters($parameters) as $parameter => $value) {
@@ -201,8 +200,7 @@ class TextDescriptor extends Descriptor
         $tagsCount = count($maxTags);
         $tagsNames = array_keys($maxTags);
 
-        $table = new Table($this->getOutput());
-        $table->setStyle('compact');
+        $table = new Table($this->output);
         $table->setHeaders(array_merge(array('Service ID'), $tagsNames, array('Class name')));
 
         foreach ($this->sortServiceIds($serviceIds) as $serviceId) {
@@ -337,11 +335,11 @@ class TextDescriptor extends Descriptor
         $this->writeText($this->formatSection('event_dispatcher', $label)."\n", $options);
 
         $registeredListeners = $eventDispatcher->getListeners($event);
+        $table = new Table($this->output);
 
         if (null !== $event) {
             $this->writeText("\n");
-            $table = new Table($this->getOutput());
-            $table->getStyle()->setCellHeaderFormat('%s');
+
             $table->setHeaders(array('Order', 'Callable'));
 
             foreach ($registeredListeners as $order => $listener) {
@@ -354,8 +352,6 @@ class TextDescriptor extends Descriptor
             foreach ($registeredListeners as $eventListened => $eventListeners) {
                 $this->writeText(sprintf("\n<info>[Event]</info> %s\n", $eventListened), $options);
 
-                $table = new Table($this->getOutput());
-                $table->getStyle()->setCellHeaderFormat('%s');
                 $table->setHeaders(array('Order', 'Callable'));
 
                 foreach ($eventListeners as $order => $eventListener) {