+++ src/Symfony/Bundle/TwigBundle/Tests/Controller/PreviewErrorControllerTest.php
@@ -21,8 +21,6 @@ class PreviewErrorControllerTest extends TestCase
 {
     public function testForwardRequestToConfiguredController()
     {
-        $self = $this;
-
         $request = Request::create('whatever');
         $response = new Response("");
         $code = 123;
@@ -33,15 +31,14 @@ class PreviewErrorControllerTest extends TestCase
             ->expects($this->once())
             ->method('handle')
             ->with(
-                $this->callback(function (Request $request) use ($self, $logicalControllerName, $code) {
+                $this->callback(function (Request $request) use ($logicalControllerName, $code) {
 
-                    $self->assertEquals($logicalControllerName, $request->attributes->get('_controller'));
+                    $this->assertEquals($logicalControllerName, $request->attributes->get('_controller'));
 
                     $exception = $request->attributes->get('exception');
-                    $self->assertInstanceOf('Symfony\Component\Debug\Exception\FlattenException', $exception);
-                    $self->assertEquals($code, $exception->getStatusCode());
-
-                    $self->assertFalse($request->attributes->get('showException'));
+                    $this->assertInstanceOf('Symfony\Component\HttpKernel\Exception\FlattenException', $exception);
+                    $this->assertEquals($code, $exception->getStatusCode());
+                    $this->assertFalse($request->attributes->get('showException'));
 
                     return true;
                 }),