+++ src/Symfony/Component/EventDispatcher/Tests/EventTest.php
@@ -25,18 +25,12 @@ class EventTest extends \PHPUnit_Framework_TestCase
     protected $event;
 
     /**
-     * @var \Symfony\Component\EventDispatcher\EventDispatcher
-     */
-    protected $dispatcher;
-
-    /**
      * Sets up the fixture, for example, opens a network connection.
      * This method is called before a test is executed.
      */
     protected function setUp()
     {
         $this->event = new Event();
-        $this->dispatcher = new EventDispatcher();
     }
 
     /**
@@ -46,7 +40,6 @@ class EventTest extends \PHPUnit_Framework_TestCase
     protected function tearDown()
     {
         $this->event = null;
-        $this->dispatcher = null;
     }
 
     public function testIsPropagationStopped()
@@ -58,31 +51,5 @@ class EventTest extends \PHPUnit_Framework_TestCase
     {
         $this->event->stopPropagation();
         $this->assertTrue($this->event->isPropagationStopped());
-    }
-
-    public function testLegacySetDispatcher()
-    {
-        $this->iniSet('error_reporting', -1 & ~E_USER_DEPRECATED);
-        $this->event->setDispatcher($this->dispatcher);
-        $this->assertSame($this->dispatcher, $this->event->getDispatcher());
-    }
-
-    public function testLegacyGetDispatcher()
-    {
-        $this->iniSet('error_reporting', -1 & ~E_USER_DEPRECATED);
-        $this->assertNull($this->event->getDispatcher());
-    }
-
-    public function testLegacyGetName()
-    {
-        $this->iniSet('error_reporting', -1 & ~E_USER_DEPRECATED);
-        $this->assertNull($this->event->getName());
-    }
-
-    public function testLegacySetName()
-    {
-        $this->iniSet('error_reporting', -1 & ~E_USER_DEPRECATED);
-        $this->event->setName('foo');
-        $this->assertEquals('foo', $this->event->getName());
     }
 }