+++ src/Symfony/Component/Form/Tests/ResolvedFormTypeTest.php
@@ -56,13 +56,11 @@ class ResolvedFormTypeTest extends \PHPUnit_Framework_TestCase
 
     public function testGetOptionsResolver()
     {
-        $test = $this;
         $i = 0;
 
-        $assertIndexAndAddOption = function ($index, $option, $default) use (&$i, $test) {
-            return function (OptionsResolver $resolver) use (&$i, $test, $index, $option, $default) {
-                /* @var \PHPUnit_Framework_TestCase $test */
-                $test->assertEquals($index, $i, 'Executed at index '.$index);
+        $assertIndexAndAddOption = function ($index, $option, $default) use (&$i) {
+            return function (OptionsResolver $resolver) use (&$i, $index, $option, $default) {
+                $this->assertEquals($index, $i, 'Executed at index '.$index);
 
                 ++$i;
 
@@ -155,13 +153,11 @@ class ResolvedFormTypeTest extends \PHPUnit_Framework_TestCase
 
     public function testBuildForm()
     {
-        $test = $this;
         $i = 0;
 
-        $assertIndex = function ($index) use (&$i, $test) {
-            return function () use (&$i, $test, $index) {
-                /* @var \PHPUnit_Framework_TestCase $test */
-                $test->assertEquals($index, $i, 'Executed at index '.$index);
+        $assertIndex = function ($index) use (&$i) {
+            return function () use (&$i, $index) {
+                $this->assertEquals($index, $i, 'Executed at index '.$index);
 
                 ++$i;
             };
@@ -223,13 +219,11 @@ class ResolvedFormTypeTest extends \PHPUnit_Framework_TestCase
         $form = $this->getMock('Symfony\Component\Form\Test\FormInterface');
         $view = $this->getMock('Symfony\Component\Form\FormView');
 
-        $test = $this;
         $i = 0;
 
-        $assertIndex = function ($index) use (&$i, $test) {
-            return function () use (&$i, $test, $index) {
-                /* @var \PHPUnit_Framework_TestCase $test */
-                $test->assertEquals($index, $i, 'Executed at index '.$index);
+        $assertIndex = function ($index) use (&$i) {
+            return function () use (&$i, $index) {
+                $this->assertEquals($index, $i, 'Executed at index '.$index);
 
                 ++$i;
             };
@@ -267,13 +261,11 @@ class ResolvedFormTypeTest extends \PHPUnit_Framework_TestCase
         $form = $this->getMock('Symfony\Component\Form\Test\FormInterface');
         $view = $this->getMock('Symfony\Component\Form\FormView');
 
-        $test = $this;
         $i = 0;
 
-        $assertIndex = function ($index) use (&$i, $test) {
-            return function () use (&$i, $test, $index) {
-                /* @var \PHPUnit_Framework_TestCase $test */
-                $test->assertEquals($index, $i, 'Executed at index '.$index);
+        $assertIndex = function ($index) use (&$i) {
+            return function () use (&$i, $index) {
+                $this->assertEquals($index, $i, 'Executed at index '.$index);
 
                 ++$i;
             };