+++ src/Symfony/Component/OptionsResolver/Tests/LegacyOptionsResolverTest.php
@@ -113,16 +113,13 @@ class LegacyOptionsResolverTest extends \PHPUnit_Framework_TestCase
 
     public function testResolveLazyDependencyOnMissingOptionalWithoutDefault()
     {
-        $test = $this;
-
         $this->resolver->setOptional(array(
             'one',
         ));
 
         $this->resolver->setDefaults(array(
-            'two' => function (Options $options) use ($test) {
-                /* @var \PHPUnit_Framework_TestCase $test */
-                $test->assertFalse(isset($options['one']));
+            'two' => function (Options $options) {
+                $this->assertFalse(isset($options['one']));
 
                 return '2';
             },
@@ -137,16 +134,13 @@ class LegacyOptionsResolverTest extends \PHPUnit_Framework_TestCase
 
     public function testResolveLazyDependencyOnOptionalWithoutDefault()
     {
-        $test = $this;
-
         $this->resolver->setOptional(array(
             'one',
         ));
 
         $this->resolver->setDefaults(array(
-            'two' => function (Options $options) use ($test) {
-                /* @var \PHPUnit_Framework_TestCase $test */
-                $test->assertTrue(isset($options['one']));
+            'two' => function (Options $options) {
+                $this->assertTrue(isset($options['one']));
 
                 return $options['one'].'2';
             },
@@ -185,12 +179,9 @@ class LegacyOptionsResolverTest extends \PHPUnit_Framework_TestCase
 
     public function testResolveLazyReplaceDefaults()
     {
-        $test = $this;
-
         $this->resolver->setDefaults(array(
-            'one' => function (Options $options) use ($test) {
-                /* @var \PHPUnit_Framework_TestCase $test */
-                $test->fail('Previous closure should not be executed');
+            'one' => function (Options $options) {
+                $this->fail('Previous closure should not be executed');
             },
         ));