+++ src/Symfony/Component/Translation/Tests/Dumper/JsonFileDumperTest.php
@@ -18,10 +18,6 @@ class JsonFileDumperTest extends \PHPUnit_Framework_TestCase
 {
     public function testDump()
     {
-        if (PHP_VERSION_ID < 50400) {
-            $this->markTestIncomplete('PHP below 5.4 doesn\'t support JSON pretty printing');
-        }
-
         $catalogue = new MessageCatalogue('en');
         $catalogue->add(array('foo' => 'bar'));