+++ src/Symfony/Component/VarDumper/Tests/HtmlDumperTest.php
@@ -40,7 +40,7 @@ class HtmlDumperTest extends \PHPUnit_Framework_TestCase
         ob_start();
         $dumper->dump($data);
         $out = ob_get_clean();
-        $closureLabel = PHP_VERSION_ID >= 50400 ? 'public method' : 'function';
+        $closureLabel = 'public method';
         $out = preg_replace('/[ \t]+$/m', '', $out);
         $var['file'] = htmlspecialchars($var['file'], ENT_QUOTES, 'UTF-8');
         $intMax = PHP_INT_MAX;